choose photo product depending on the recipient country
Depending on the recipient country and thus the printing country we
have to choose different printing products/size.
This is a bit hackish way to choose for GB, US, etc. 4x6 as  product
for the other countries we use 10x15_cm or - depending on the photo
size '9x13_cm'.
This is actually only needed for the UK right now, because the UK is
the only country with international shipping. And we need this for an
order to Slovakia.