for international shipping we must choose the "Pro" qualityLevel
only the GB printing facility can ship internationally.
"internationally" means that we want to send to a country that has no
printing facilities.
also only Pro prints can be shipped internationally.

this will not work for us in the future because the Pro prints are way
to expensive. probably we will need to limit the countries we deliver
to, to the countries that have pwinty printing facilities.