optimized email processed notification and create oder AFTER the incoming email is processed
The order is now generated after the email has been processed. Which
means it could be the case that an order will have more photos as set
in the photo limit. This is better because it does include all pictures
of a user and not only part of it. Also the code will be nicer.

The notification now also includes information about if a new order has
created and it informs the user about potential bad quality of his
photos.
Also we store which user has created the oder.
i18n is missing for the email.